Read filtering state on load
Since filtering state is saved between reloads, it makes sense to set
the filters status according to the current filters state.

This fix reads the state of the filter at load time and sets the buttons
accordingly.  If all pet types are currently not filtered (i.e. all pet
types are showing) we update the ui to not show any filter.